+++ gnpy/core/network.py
@@ -16,12 +16,13 @@ from logging import getLogger
 from os import path
 from operator import itemgetter, attrgetter
 from gnpy.core import elements
-from gnpy.core.elements import Fiber, Edfa, Transceiver, Roadm, Fused
+from gnpy.core.elements import Fiber, Edfa, Transceiver, Roadm, Fused, RamanFiber
 from gnpy.core.equipment import edfa_nf
 from gnpy.core.exceptions import ConfigurationError, NetworkTopologyError
 from gnpy.core.units import UNITS
 from gnpy.core.utils import (load_json, save_json, round2float, db2lin,
                             merge_amplifier_restrictions)
+from gnpy.core.science_utils import SimParams
 from collections import namedtuple
 
 logger = getLogger(__name__)
@@ -548,3 +549,12 @@ def build_network(network, equipment, pref_ch_db, pref_total_db):
         trx = [t for t in network.nodes() if isinstance(t, Transceiver)]
         for t in trx:
             set_egress_amplifier(network, t, equipment, pref_total_db)
+
+def load_sim_params(filename):
+    sim_params = load_json(filename)
+    return SimParams(params=sim_params)
+
+def configure_network(network, sim_params):
+    for node in network.nodes:
+        if isinstance(node, RamanFiber):
+            node.sim_params = sim_params
